requirements.txt
    langchain-community>=0.3.14
    langchain>=0.3.14
    numpy>=1.26.4
    openai>=1.59.3
    pandas>=2.2.3
    plotly>=5.24.1
    streamlit>=1.41.1
    langchain-core>=0.3.29
    faiss-cpu>=1.9.0.post1
    sentence-transformers>=3.3.1
    langchain-openai>=0.2.14